tests/e2e/requirements.txt
pytest==8.3.5
paho-mqtt==2.1.0
requests==2.32.3
